Nexus: Move `PseudoSet` and related code to `pseudoset.py`, move its tests to `test_pseudoset.py` as well.
