+++ nexus/nexus/rmg.py
@@ -4,7 +4,7 @@
 
 
 from .simulation import Simulation
-from .pseudopotential import PseudoSet
+from .pseudoset import PseudoSet
 from .rmg_input import RmgInput, generate_rmg_input
 from .rmg_analyzer import RmgAnalyzer
