+++ nexus/nexus/tests/test_vasp_input.py
@@ -4,7 +4,7 @@ pytestmark = pytest.mark.order(NexusTestOrder.VASP_ANALYZER)
 
 
 from ..generic import NexusError
-from ..pseudopotential import PseudoSet
+from ..pseudoset import PseudoSet
 
 from nexus.nexus_base import nexus_core
 from . import isolate_nexus_core, TEST_DIR
