+++ nexus/nexus/vasp_input.py
@@ -23,7 +23,7 @@ from .nexus_base import nexus_noncore
 from .simulation import SimulationInput
 from .structure import interpolate_structures, Structure
 from .physical_system import PhysicalSystem
-from .pseudopotential import PseudoSet
+from .pseudoset import PseudoSet
 from .developer import DevBase, obj, error
 from .utilities import path_string
 from . import numpy_extensions as npe
